refactor usecase
