+++ packages/create-llama/helpers/types.ts
@@ -49,23 +49,6 @@ export type TemplateUseCase =
   | "document_generator"
   | "hitl";
 
-export const ALL_TYPESCRIPT_USE_CASES: TemplateUseCase[] = [
-  "agentic_rag",
-  "deep_research",
-  "financial_report",
-  "code_generator",
-  "document_generator",
-  "hitl",
-];
-
-export const ALL_PYTHON_USE_CASES: TemplateUseCase[] = [
-  "agentic_rag",
-  "deep_research",
-  "financial_report",
-  "code_generator",
-  "document_generator",
-];
-
 // Config for both file and folder
 export type FileSourceConfig =
   | {
